保存した学習記録の内容を閲覧できるようにしました

+++ work/app/functions.php
@@ -34,11 +34,11 @@ function validateToken(){
 function addTodo($pdo){
     /* ------------------------------------サーバに送られてきたデータを取得----------------------------- */
 $pro_summary = filter_input(INPUT_POST,'pro_summary'); /* 問題概要の内容 */
-// $pro_detail = filter_input(INPUT_POST,'pro_detail'); /* 問題詳細の内容 */
-// $pro_attachment = filter_input(INPUT_POST,'pro_attachment'); /* 問題の添付ファイルの内容 */
-// $pro_sosummary = filter_input(INPUT_POST,'pro_sosummary'); /* 対処概要の内容 */
-// $pro_sodetail = filter_input(INPUT_POST,'pro_sodetail'); /* 対処詳細の内容 */
-// $so_attachment = filter_input(INPUT_POST,'so_attachment'); /* 解決の添付ファイルの内容 */
+$pro_detail = filter_input(INPUT_POST,'pro_detail'); /* 問題詳細の内容 */
+$pro_attachment = filter_input(INPUT_POST,'pro_attachment'); /* 問題の添付ファイルの内容 */
+$pro_sosummary = filter_input(INPUT_POST,'pro_sosummary'); /* 対処概要の内容 */
+$pro_sodetail = filter_input(INPUT_POST,'pro_sodetail'); /* 対処詳細の内容 */
+$so_attachment = filter_input(INPUT_POST,'so_attachment'); /* 解決の添付ファイルの内容 */
 /* ---------------------------------------------ここまで--------------------------------------- */
 
 $status = filter_input(INPUT_POST,'status');
@@ -46,13 +46,25 @@ $status = filter_input(INPUT_POST,'status');
     /* -------------------------------データベースへ送信------------------------------------------- */
     switch ($status){
         case 'done':  /* 完了ボタンの場合 */
-            $stmt = $pdo->prepare("INSERT INTO worklists (title) VALUES (:pro_summary)");
+            $stmt = $pdo->prepare("INSERT INTO worklists (pro_summary,proDetail,proAttachment,soSummary,soDetail,soAttachment,folderNo) VALUES (:pro_summary,:proDetail,:proAttachment,:soSummary,:soDetail,:soAttachment,:folderNo)"); /* フォルダナンバーをテーブルに挿入 */
             $stmt->bindValue('pro_summary', $pro_summary, PDO::PARAM_STR);
+            $stmt->bindValue('proDetail', $pro_detail, PDO::PARAM_STR);
+            $stmt->bindValue('proAttachment', $pro_attachment, PDO::PARAM_STR); //添付ファイルなのであとで型を変更する
+            $stmt->bindValue('soSummary',$pro_sosummary, PDO::PARAM_STR);
+            $stmt->bindValue('soDetail', $pro_sodetail, PDO::PARAM_STR);
+            $stmt->bindValue('soAttachment',$so_attachment, PDO::PARAM_STR);
+            $stmt->bindValue('folderNo', $_SESSION['folderNo'], PDO::PARAM_INT);
             $stmt->execute();
             break;
         case 'notDone': /* 未完了ボタンの場合 */
-            $stmt = $pdo->prepare("INSERT INTO worklists (title,is_done) VALUES (:pro_summary,0)");
+            $stmt = $pdo->prepare(" INSERT INTO worklists (pro_summary,proDetail,proAttachment,soSummary,soDetail,soAttachment,is_done,folderNo) VALUES (:pro_summary,:proDetail,:proAttachment,:soSummary,:soDetail,:soAttachment,0,:folderNo)"); /* フォルダナンバーをテーブルに挿入 */
             $stmt->bindValue('pro_summary', $pro_summary, PDO::PARAM_STR);
+            $stmt->bindValue('proDetail', $pro_detail, PDO::PARAM_STR);
+            $stmt->bindValue('proAttachment', $pro_attachment, PDO::PARAM_STR); //添付ファイルなのであとで型を変更する
+            $stmt->bindValue('soSummary',$pro_sosummary, PDO::PARAM_STR);
+            $stmt->bindValue('soDetail', $pro_sodetail, PDO::PARAM_STR);
+            $stmt->bindValue('soAttachment',$so_attachment, PDO::PARAM_STR);
+            $stmt->bindValue('folderNo', $_SESSION['folderNo'], PDO::PARAM_INT);
             $stmt->execute();
             break;
     }
@@ -61,11 +73,40 @@ $status = filter_input(INPUT_POST,'status');
 
 function deleteTodo($pdo){
     $id = filter_input(INPUT_POST,'id');
-
+    if(empty($id)){
+        return;
+    }
     $stmt = $pdo->prepare("DELETE FROM worklists WHERE id = :id");
     $stmt->bindValue('id',$id,PDO::PARAM_INT);
     $stmt->execute();
 
 }
+
+
+function getTodo($pdo){
+    $folderNo = $_COOKIE['folderNo'];
+    $stmt = $pdo->prepare("SELECT * FROM worklists WHERE folderNo = :folderNo");
+    $stmt->execute(['folderNo'=>$folderNo]);
+    $worklists = $stmt->fetchAll();
+    return $worklists;
+}
+
+function createFolders(){
+    $icon = filter_input(INPUT_POST,'icon');/* <i class='fas fa-folder fa-3x'></i>を受け取る */
+    $fp = fopen(FILENAME,'a'); /* テキストファイルを追記モードで開く */
+    fwrite($fp,$icon.PHP_EOL); /* テキストファイルに追記*/
+    fclose($fp);
+    header('Location: http://localhost:8562/top.php');
+    exit;
+}
+
+function getFolderNo(){
+
+    $folderNo = filter_input(INPUT_POST,'folderNo');
+    setcookie('folderNo',$folderNo);
+    return $folderNo;
+
+}
+    
 
 ?>
